+++ integration-tests/api/__tests__/batch-jobs/product/import.js
@@ -183,7 +183,7 @@ describe("Product import batch job", () => {
               ean: null,
               upc: null,
               inventory_quantity: 10,
-              prices: [
+              prices: expect.arrayContaining([
                 expect.objectContaining({
                   currency_code: "eur",
                   amount: 100,
@@ -198,7 +198,7 @@ describe("Product import batch job", () => {
                   amount: 130,
                   region_id: "region-product-import-1",
                 }),
-              ],
+              ]),
               options: expect.arrayContaining([
                 expect.objectContaining({
                   value: "option 1 value red",
@@ -210,24 +210,24 @@ describe("Product import batch job", () => {
             }),
           ],
           type: null,
-          images: [
+          images: expect.arrayContaining([
             expect.objectContaining({
               url: "test-image.png",
             }),
-          ],
-          options: [
+          ]),
+          options: expect.arrayContaining([
             expect.objectContaining({
               title: "test-option-1",
             }),
             expect.objectContaining({
               title: "test-option-2",
             }),
-          ],
-          tags: [
+          ]),
+          tags: expect.arrayContaining([
             expect.objectContaining({
               value: "123_1",
             }),
-          ],
+          ]),
           collection: expect.objectContaining({
             handle: collectionHandle1,
           }),
@@ -249,7 +249,7 @@ describe("Product import batch job", () => {
               ean: null,
               upc: null,
               inventory_quantity: 10,
-              prices: [
+              prices: expect.arrayContaining([
                 expect.objectContaining({
                   currency_code: "eur",
                   amount: 100,
@@ -264,7 +264,7 @@ describe("Product import batch job", () => {
                   amount: 130,
                   region_id: "region-product-import-1",
                 }),
-              ],
+              ]),
               options: expect.arrayContaining([
                 expect.objectContaining({
                   value: "option 1 value red",
@@ -276,19 +276,19 @@ describe("Product import batch job", () => {
             }),
           ],
           type: null,
-          images: [
+          images: expect.arrayContaining([
             expect.objectContaining({
               url: "test-image.png",
             }),
-          ],
-          options: [
+          ]),
+          options: expect.arrayContaining([
             expect.objectContaining({
               title: "test-option-1",
             }),
             expect.objectContaining({
               title: "test-option-2",
             }),
-          ],
+          ]),
           tags: [],
           collection: expect.objectContaining({
             handle: collectionHandle1,