+++ packages/admin-ui/ui/src/components/templates/inventory-table/index.tsx
@@ -47,7 +47,7 @@ const LocationDropdown = ({
   const { stock_locations: locations, isLoading } = useAdminStockLocations()
 
   useEffect(() => {
-    if (!selectedLocation && !isLoading && locations) {
+    if (!selectedLocation && !isLoading && locations?.length) {
       onChange(locations[0].id)
     }
   }, [isLoading, locations, onChange, selectedLocation])
@@ -85,6 +85,9 @@ const InventoryTable: React.FC<InventoryTableProps> = () => {
 
   const location = useLocation()
 
+  const { stock_locations, isLoading: locationsLoading } =
+    useAdminStockLocations()
+
   const defaultQuery = useMemo(() => {
     if (store) {
       return {
@@ -96,16 +99,9 @@ const InventoryTable: React.FC<InventoryTableProps> = () => {
   }, [store])
 
   const {
-    removeTab,
-    setTab,
-    saveTab,
-    availableTabs: filterTabs,
-    activeFilterTab,
     reset,
     paginate,
-    setFilters,
     setLocationFilter,
-    filters,
     setQuery: setFreeText,
     queryObject,
     representationObject,
@@ -116,11 +112,6 @@ const InventoryTable: React.FC<InventoryTableProps> = () => {
 
   const [query, setQuery] = useState(queryObject.query)
   const [numPages, setNumPages] = useState(0)
-
-  const clearFilters = () => {
-    reset()
-    setQuery("")
-  }
 
   const { inventory_items, isLoading, count } = useAdminInventoryItems(
     {
@@ -284,6 +275,11 @@ const InventoryTable: React.FC<InventoryTableProps> = () => {
           })}
         </Table.Body>
       </Table>
+      {!rows.length && !locationsLoading && !stock_locations?.length && (
+        <div className="text-grey-50 w-full py-8 text-center">
+          You don't have any stock locations. Add one to see inventory.
+        </div>
+      )}
     </TableContainer>
   )
 }