+++ packages/medusa-payment-paypal/src/services/utils/utils.ts
@@ -1,0 +1,8 @@
+import { zeroDecimalCurrencies } from "medusa-core-utils";
+
+export function roundToTwo(num: number, currency: string): string {
+  if (zeroDecimalCurrencies.includes(currency.toLowerCase())) {
+    return `${num}`
+  }
+  return num.toFixed(2)
+}