+++ packages/medusa/src/services/discount.ts
@@ -619,12 +619,9 @@ class DiscountService extends TransactionBaseService {
         })
 
       let fullItemPrice = lineItem.unit_price * lineItem.quantity
-      if (
-        this.featureFlagRouter_.isFeatureEnabled(
-          TaxInclusivePricingFeatureFlag.key
-        ) &&
-        lineItem.includes_tax
-      ) {
+      const includesTax = this.featureFlagRouter_.isFeatureEnabled(TaxInclusivePricingFeatureFlag.key) && lineItem.includes_tax
+
+      if (includesTax) {
         const lineItemTotals = await this.newTotalsService_
           .withTransaction(transactionManager)
           .getLineItemTotals([lineItem], {
@@ -649,6 +646,7 @@ class DiscountService extends TransactionBaseService {
         const totals = await this.newTotalsService_.getLineItemTotals(
           discountedItems,
           {
+            includeTax: includesTax,
             calculationContext,
           }
         )