+++ packages/types/src/stock-location/stock-location.ts
@@ -1,4 +1,3 @@
-import { FindConfig } from "../common/common"
 import {
   CreateStockLocationInput,
   FilterableStockLocationProps,
@@ -6,25 +5,38 @@ import {
   UpdateStockLocationInput,
 } from "./common"
 
+import { FindConfig } from "../common/common"
+import { SharedContext } from "../shared-context"
+
 export interface IStockLocationService {
   list(
     selector: FilterableStockLocationProps,
-    config?: FindConfig<StockLocationDTO>
+    config?: FindConfig<StockLocationDTO>,
+    context?: SharedContext
   ): Promise<StockLocationDTO[]>
 
   listAndCount(
     selector: FilterableStockLocationProps,
-    config?: FindConfig<StockLocationDTO>
+    config?: FindConfig<StockLocationDTO>,
+    context?: SharedContext
   ): Promise<[StockLocationDTO[], number]>
 
   retrieve(
     id: string,
-    config?: FindConfig<StockLocationDTO>
+    config?: FindConfig<StockLocationDTO>,
+    context?: SharedContext
   ): Promise<StockLocationDTO>
 
-  create(input: CreateStockLocationInput): Promise<StockLocationDTO>
+  create(
+    input: CreateStockLocationInput,
+    context?: SharedContext
+  ): Promise<StockLocationDTO>
 
-  update(id: string, input: UpdateStockLocationInput): Promise<StockLocationDTO>
+  update(
+    id: string,
+    input: UpdateStockLocationInput,
+    context?: SharedContext
+  ): Promise<StockLocationDTO>
 
-  delete(id: string): Promise<void>
+  delete(id: string, context?: SharedContext): Promise<void>
 }